# Restructure qdigest serialization