--- a/stats/src/test/java/io/airlift/stats/BenchmarkQuantileDigest.java
+++ b/stats/src/test/java/io/airlift/stats/BenchmarkQuantileDigest.java
@@ -1,6 +1,7 @@
 package io.airlift.stats;
 
 import com.google.common.collect.ImmutableList;
+import io.airlift.slice.Slice;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.Fork;
 import org.openjdk.jmh.annotations.Measurement;
@@ -60,12 +61,14 @@ public class BenchmarkQuantileDigest
     {
         private QuantileDigest digest1;
         private QuantileDigest digest2;
+        private Slice serializedDigest;
 
         @Setup
         public void setup(Data data)
         {
             digest1 = makeDigest(data.values1);
             digest2 = makeDigest(data.values2);
+            serializedDigest = digest1.serialize();
         }
 
         private QuantileDigest makeDigest(long[] values)
@@ -103,6 +106,18 @@ public class BenchmarkQuantileDigest
         QuantileDigest merged = new QuantileDigest(data.digest1);
         merged.merge(data.digest2);
         return merged;
+    }
+
+    @Benchmark
+    public QuantileDigest benchmarkDeserialize(Digest data)
+    {
+        return new QuantileDigest(data.serializedDigest);
+    }
+
+    @Benchmark
+    public Slice benchmarkSerialize(Digest data)
+    {
+        return data.digest1.serialize();
     }
 
     @Benchmark