… stats/docs/qdigest.graffle
… stats/docs/qdigest_layout.png
… stats/docs/qdigest_node_layout.png
… stats/docs/qdigest_node_structure_byte.png