* Instead of having 4 groups of post-order traversed sequences, have
  one sequence which contains this information (count, level, value,
  left/right child)
* Pack the bits between level and children since we can express this
  in just one byte.  This saves one byte per node in the tree (a roughly 10% savings in size).
* Add a format byte at the beginning to allow for the layout to 
  change shape in the future.

CC: @martint