@tdcmeehan 2018-10-18
I've also added a benchmark.  Here are the results for serializing/deserializing.


New:
```
Benchmark                                                                       Mode  Cnt       Score      Error   Units
BenchmarkQuantileDigest.benchmarkDeserialize                                   thrpt   30   23349.172 ±  906.412   ops/s
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.alloc.rate                    thrpt   30    1847.956 ±   83.224  MB/sec
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.alloc.rate.norm               thrpt   30  125037.409 ± 7522.545    B/op
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.churn.PS_Eden_Space           thrpt   30    1859.703 ±   96.126  MB/sec
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.churn.PS_Eden_Space.norm      thrpt   30  125835.097 ± 8230.996    B/op
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.churn.PS_Survivor_Space       thrpt   30       0.262 ±    0.053  MB/sec
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.churn.PS_Survivor_Space.norm  thrpt   30      17.822 ±    3.824    B/op
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.count                         thrpt   30     416.000             counts
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.time                          thrpt   30     277.000                 ms
BenchmarkQuantileDigest.benchmarkSerialize                                     thrpt   30   10701.142 ±  176.090   ops/s
BenchmarkQuantileDigest.benchmarkSerialize:·gc.alloc.rate                      thrpt   30     616.398 ±   13.714  MB/sec
BenchmarkQuantileDigest.benchmarkSerialize:·gc.alloc.rate.norm                 thrpt   30   90720.166 ± 1535.688    B/op
BenchmarkQuantileDigest.benchmarkSerialize:·gc.churn.PS_Eden_Space             thrpt   30     623.704 ±   24.694  MB/sec
BenchmarkQuantileDigest.benchmarkSerialize:·gc.churn.PS_Eden_Space.norm        thrpt   30   91786.267 ± 3240.187    B/op
BenchmarkQuantileDigest.benchmarkSerialize:·gc.churn.PS_Survivor_Space         thrpt   30       0.200 ±    0.050  MB/sec
BenchmarkQuantileDigest.benchmarkSerialize:·gc.churn.PS_Survivor_Space.norm    thrpt   30      29.379 ±    7.311    B/op
BenchmarkQuantileDigest.benchmarkSerialize:·gc.count                           thrpt   30     347.000             counts
BenchmarkQuantileDigest.benchmarkSerialize:·gc.time                            thrpt   30     230.000                 ms
```

Old: 

```
Benchmark                                                                       Mode  Cnt       Score      Error   Units
BenchmarkQuantileDigest.benchmarkDeserialize                                   thrpt   30   18304.110 ±  611.182   ops/s
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.alloc.rate                    thrpt   30    1541.509 ±   47.375  MB/sec
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.alloc.rate.norm               thrpt   30  133181.430 ± 7933.132    B/op
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.churn.PS_Eden_Space           thrpt   30    1553.405 ±   69.698  MB/sec
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.churn.PS_Eden_Space.norm      thrpt   30  134219.107 ± 9185.472    B/op
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.churn.PS_Survivor_Space       thrpt   30       0.281 ±    0.059  MB/sec
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.churn.PS_Survivor_Space.norm  thrpt   30      24.304 ±    5.396    B/op
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.count                         thrpt   30     400.000             counts
BenchmarkQuantileDigest.benchmarkDeserialize:·gc.time                          thrpt   30     268.000                 ms
BenchmarkQuantileDigest.benchmarkSerialize                                     thrpt   30   10167.300 ±  156.795   ops/s
BenchmarkQuantileDigest.benchmarkSerialize:·gc.alloc.rate                      thrpt   30     569.640 ±    9.917  MB/sec
BenchmarkQuantileDigest.benchmarkSerialize:·gc.alloc.rate.norm                 thrpt   30   88221.512 ±  276.922    B/op
BenchmarkQuantileDigest.benchmarkSerialize:·gc.churn.PS_Eden_Space             thrpt   30     578.084 ±   23.018  MB/sec
BenchmarkQuantileDigest.benchmarkSerialize:·gc.churn.PS_Eden_Space.norm        thrpt   30   89538.200 ± 3285.073    B/op
BenchmarkQuantileDigest.benchmarkSerialize:·gc.churn.PS_Survivor_Space         thrpt   30       0.188 ±    0.056  MB/sec
BenchmarkQuantileDigest.benchmarkSerialize:·gc.churn.PS_Survivor_Space.norm    thrpt   30      29.092 ±    8.698    B/op
BenchmarkQuantileDigest.benchmarkSerialize:·gc.count                           thrpt   30     317.000             counts
BenchmarkQuantileDigest.benchmarkSerialize:·gc.time                            thrpt   30     208.000                 ms
```