@tdcmeehan 2018-11-05
Thanks @martint.  I addressed your comment and added the documentation on the serialization format.  Can you please give it another pass.