@tdcmeehan 2018-11-06
Thanks for the review @martint.  I addressed those points.